refactor(agents): extract CodeToolAuthDelegate from AgentEditorViewModel
Move the Code Interpreter (execute_code) tool-key flow into
CodeToolAuthDelegate: auth verification, key install/revoke, and the
auth-gated capability toggle. The VM exposes the same public methods as thin
delegations and runs the initial verify through the delegate.

No behavior change.
